Renamed WebSocket.h to BeastWebSocketResource.h.
