+++ vnext/Desktop/BeastWebSocketResource.h
