+++ vnext/Desktop/React.Windows.Desktop.vcxproj.filters
@@ -161,7 +161,7 @@
     <ClInclude Include="NativeModuleFactories.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="WebSocket.h">
+    <ClInclude Include="BeastWebSocketResource.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
